+++ lib/src/risk_system/pricing/calculators/elec_option/leaf.dart
@@ -18,7 +18,7 @@ class LeafElecOption extends LeafBase {
     required this.underlyingPrice,
     required this.volatility,
     required this.fixPrice,
-  }) : expirationDate = lastBusinessDayPrior(month) {
+  }) : expirationDate = lastBusinessDayPrior(month.startDate) {
     _bs = BlackScholes(
         type: callPut,
         strike: strike,