+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Release 2024-03-05
 - Add Cal 1x option expiration date, in lastTradingDayForCalendar1xOptions()
+- Refactor lastBusinessDayPrior() to take a date not a month argument.  
 
 ## Release 2024-03-04
 - Add an IceCalendar for US energy.